requirements.txt
numpy
pandas
python-dotenv
Pillow
extcolors
ultralytics